requirements.txt
mysql-connector-python==9.4.0
pandas==2.3.2
Flask==3.0.0
Werkzeug==3.0.1
certifi==2025.10.5
charset-normalizer==3.4.3
cloudinary==1.44.1
dropbox==12.0.2
idna==3.10
ply==3.11
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
stone==3.3.1
urllib3==2.5.0
gunicorn==21.2.0 
